# Replace some distance checks with square distance checks in NavMap