Replaced some distance checks with square distance checks in NavMap, wherever the purpose is only to find the nearest element. Some of these checks can be made thousands of times a frame for path finding, so this should cut down on some math costs. Places where the distance is actually used in some sort of scoring formula were left alone.