@smix8 2024-06-13
Would like to see some performance tests, the change looks sound on paper.

Tried with this PR and master and to my surprise I get 2-3 fps on average less with 1000 agents on an unoptimized dev build with this PR and I have no explanation why.