@MajorMcDoom 2024-07-02
I just did a test with this project:
[nav-sqr-dist-test.zip](https://github.com/user-attachments/files/16074199/nav-sqr-dist-test.zip)

Results:
```
[BIG MAP, SINGLE REGION] -------------------------------------------------------

# WITHOUT sqr_dist
map_get_closest_point_to_segment (intersecting): 2547.928 ms
map_get_closest_point_to_segment (non-intersecting): 3510.255 ms
map_get_closest_point_to_segment (intersecting collision): 672.153 ms
map_get_closest_point_to_segment (non-intersecting collision): 467.859 ms
map_get_path (non-optimized): 5577.937 ms
map_get_path (optimized): 5565.387 ms

# WITH sqr_dist
map_get_closest_point_to_segment (intersecting): 2445.945 ms
map_get_closest_point_to_segment (non-intersecting): 3373.6 ms
map_get_closest_point_to_segment (intersecting collision): 659.314 ms
map_get_closest_point_to_segment (non-intersecting collision): 444.546 ms
map_get_path (non-optimized): 5444.421 ms
map_get_path (optimized): 5470.347 ms

[BIG MAP, MULTI REGION] --------------------------------------------------------

# WITHOUT sqr_dist
map_get_closest_point_to_segment (intersecting): 2247.26 ms
map_get_closest_point_to_segment (non-intersecting): 2892.012 ms
map_get_closest_point_to_segment (intersecting collision): 545.571 ms
map_get_closest_point_to_segment (non-intersecting collision): 377.858 ms
map_get_path (non-optimized): 2967.82 ms
map_get_path (optimized): 3012.839 ms

# WITH sqr_dist
map_get_closest_point_to_segment (intersecting): 2186.209 ms
map_get_closest_point_to_segment (non-intersecting): 2794.83 ms
map_get_closest_point_to_segment (intersecting collision): 547.577 ms
map_get_closest_point_to_segment (non-intersecting collision): 365.616 ms
map_get_path (non-optimized): 2918.6 ms
map_get_path (optimized): 2931.453 ms

[SMALL MAP] --------------------------------------------------------------------

# WITHOUT sqr_dist
map_get_closest_point_to_segment (intersecting): 3904.244 ms
map_get_closest_point_to_segment (non-intersecting): 3585.282 ms
map_get_closest_point_to_segment (intersecting collision): 693.023 ms
map_get_closest_point_to_segment (non-intersecting collision): 477.006 ms
map_get_path (non-optimized): 4051.911 ms
map_get_path (optimized): 4090.907 ms

# WITH sqr_dist
map_get_closest_point_to_segment (intersecting): 3789.921 ms
map_get_closest_point_to_segment (non-intersecting): 3418.402 ms
map_get_closest_point_to_segment (intersecting collision): 687.428 ms
map_get_closest_point_to_segment (non-intersecting collision): 468.693 ms
map_get_path (non-optimized): 3929.711 ms
map_get_path (optimized): 4041.489 ms
```

There definitely is _an improvement_, but it's also _tiny_.
I ran lots of tests, and while there are super rare outliers where the sqr dist performance was worse, overall it is about 2%-5% faster. This is a tiny gain, which is a little unexpected, but maybe my test isn't capturing the correct hot paths?

Based on these results, I see no reason *not* to merge this, but at the same time, it's not a super compelling reason to merge either.