@MajorMcDoom 2024-10-16
@smix8 @akien-mga 
Rebased. Please let me know if I did the conflict resolution correctly, I'm still a bit new to this aspect of Git. Thanks!